Update of checkoutProcess
Turn out it is my computer configuration that cased all these issue.

The only changes that I made on the checkoutProcess.php  it the reset of the "Number of Items in Cart". I commented the $_SESSION["NumCartItem"]=0; to unset$_SESSION["NumCartItem"] so that I will remove the 0 items in the cart ... on the nav bar once a payment is made.